feat: migrate to SQLite and complete mood tracking features
## Database Migration (DuckDB → SQLite)

**Why**: DuckDB's foreign key constraints prevent UPDATE operations on
referenced rows, breaking soft delete functionality. SQLite properly
supports this pattern and is more mature for our scale (~10K records).

### Changes
- Replace duckdb with rusqlite dependency
- Convert schema: sequences → AUTOINCREMENT, TIMESTAMP → TEXT
- Fix strftime syntax (format parameter order differs)
- Update all repository and model error types
- Add ON DELETE CASCADE support (now works correctly)

### Test Results
✅ All 66 tests passing (including previously ignored soft delete test)

## User Story 2: Mood Check-In (Complete)

### Backend (Rust)
- Registered mood commands in Tauri (log_mood, get_mood_history, etc.)
- Fixed i64 → i32 for specta TypeScript binding compatibility
- Regenerated TypeScript bindings

### Frontend (Svelte)
- MoodScaleInput: 1-5 rating selector with color coding
- ActivitySelector: Multi-select with create-new functionality
- MoodHistoryList: Display past check-ins with timestamps
- /mood route: Quick logging (<30 sec target)
- /mood/history route: Full history with stats and filtering
- Mood store for reactive state management
- Color utility constants

## User Story 3: Activity Management (Complete)

### Tests
- 16 integration tests for activity CRUD operations
- Soft delete preserves historical data (NOW PASSING with SQLite)

### Frontend
- ActivityForm: Create/edit with validation
- ActivityList: Grid display with edit/delete
- /mood/activities route: Full CRUD interface
- Soft-deleted activities show "(deleted)" badge in history

## Code Quality Fixes
- Add bindings.ts to eslint ignores (auto-generated file)
- Add keys to all {#each} blocks for proper reactivity
- Remove unused variables
- Add SvelteKit preload-data attributes to links
- Disable overly strict navigation resolution rule for internal links

+++ src-tauri/src/features/mood/commands.rs
@@ -54,7 +54,7 @@ pub async fn create_activity(
 #[tauri::command]
 #[specta::specta]
 pub async fn update_activity(
-    id: i64,
+    id: i32,
     request: UpdateActivityRequest,
     state: State<'_, AppState>,
 ) -> Result<Activity, String> {
@@ -76,7 +76,7 @@ pub async fn update_activity(
 // T108: delete_activity command
 #[tauri::command]
 #[specta::specta]
-pub async fn delete_activity(id: i64, state: State<'_, AppState>) -> Result<(), String> {
+pub async fn delete_activity(id: i32, state: State<'_, AppState>) -> Result<(), String> {
     let repo = MoodRepository::new(state.db.clone());
 
     repo.delete_activity(id)
@@ -90,7 +90,7 @@ pub async fn delete_activity(id: i64, state: State<'_, AppState>) -> Result<(), 
 // T093c: delete_mood_checkin command (cascade deletion)
 #[tauri::command]
 #[specta::specta]
-pub async fn delete_mood_checkin(id: i64, state: State<'_, AppState>) -> Result<(), String> {
+pub async fn delete_mood_checkin(id: i32, state: State<'_, AppState>) -> Result<(), String> {
     let repo = MoodRepository::new(state.db.clone());
 
     repo.delete_mood_checkin(id)
