+++ src-tauri/src/features/mood/queries.rs
@@ -30,7 +30,7 @@ pub async fn get_mood_history(
 // T082: get_mood_checkin command
 #[tauri::command]
 #[specta::specta]
-pub async fn get_mood_checkin(id: i64, state: State<'_, AppState>) -> Result<MoodCheckin, String> {
+pub async fn get_mood_checkin(id: i32, state: State<'_, AppState>) -> Result<MoodCheckin, String> {
     let repo = MoodRepository::new(state.db.clone());
 
     repo.get_mood_checkin(id)
